refactor: Separate out `storacha` API
